Coloquei a Biblioteca na pasta app, e apaguei a antiga na pasta models
