--- src/app/Blibioteca.java
@@ -1,9 +1,0 @@
-package app;
-import models.Livros;
-import  models.Usuarios;
-
-public class Blibioteca {
-    public static void main(String[] args){
-        System.out.print("Classe de blibiotecas.");
-    }
-}
