--- src/models/Biblioteca.java
@@ -1,30 +1,0 @@
-package models;
-import java.util.ArrayList;
-import java.util.Scanner;
-
-public class Biblioteca {
-        private ArrayList<Usuarios> usuariosLista;
-        public ArrayList<Livros> livrosLista;
-
-        public Biblioteca() {
-            usuariosLista = new ArrayList<>();
-            livrosLista = new ArrayList<>();
-        }
-        public void addUsuario(){
-            Scanner input = new Scanner(System.in);
-            String nome;
-            System.out.println("Digite aqui as suas caracteristicas:");
-            nome = input.next();
-            this.usuariosLista.add(new Usuarios(nome));
-        }
-        public void getUsers(){
-            if (usuariosLista.isEmpty()) {
-                System.out.println("Não há usuários cadastrados.");
-            } else {
-                for(Usuarios element: usuariosLista){
-                    System.out.println(element.getNome());
-                }
-            }
-
-        }
-    }
